Amiwis

+++ core/src/cl/makinolas/atk/actors/friend/Charmander.java
@@ -28,13 +28,13 @@ public class Charmander extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Charmander(float level){
+  public Charmander(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 16, 1);
   }
