+++ core/src/cl/makinolas/atk/actors/friend/Eevee.java
@@ -21,7 +21,7 @@ public class Eevee extends AbstractFriend {
   }
 
   @Override
-  protected void initLevel(float level) {
+  protected void initLevel(int level) {
     this.level = new Level(level);    
   }
 }
