+++ core/src/cl/makinolas/atk/actors/friend/Gastly.java
@@ -23,13 +23,13 @@ public class Gastly extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Gastly(float level){
+  public Gastly(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 25, 1);
   }
