+++ core/src/cl/makinolas/atk/actors/friend/Magnemite.java
@@ -28,13 +28,13 @@ public class Magnemite extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Magnemite(float level){
+  public Magnemite(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 22, 1);
    new Evolution(this.level, 45, 2);
