+++ core/src/cl/makinolas/atk/actors/friend/Pichu.java
@@ -28,13 +28,13 @@ public class Pichu extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Pichu(float level){
+  public Pichu(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 10, 1);
   }
