+++ core/src/cl/makinolas/atk/actors/friend/Scyther.java
@@ -28,13 +28,13 @@ public class Scyther extends AbstractFriend {
     setMaxMagic(100);
   }
   
-  public Scyther(float level){
+  public Scyther(int level){
     this();
     initLevel(level);
   }
   
   @Override
-  protected void initLevel(float level){
+  protected void initLevel(int level){
    this.level = new Level(level);
    new Evolution(this.level, 30, 1);
   }
